requirements.txt
discord.py>=2.3.0
motor>=3.1.1
python-dotenv>=1.0.0
colorama>=0.4.6
aiohttp>=3.8.0
Pillow>=9.5.0
psutil>=5.9.0
